re-organised all main features in its own folder and added its respective hooks and methods
