+++ app/src/ranking/components/RankingPage.tsx
